share the move-to-folder composable across collections
